Sprint-8 Completed (#27)
* PS-105: F_Admin is not able to update a US into party view (#23)

* new changes

* url's improvement

* css change

* ccs and env improvements

* PS-102/107: F_Deny access to party view if it does not exist (#24)

* some changes

* ux improvement

* code cleanning

* code claaning 2

* PS-91: F_Create PartyPlayer view 1.0 (#25)

* new Nav-Bar behavior

* first version player view

* some changes

* PS-112: F_Update navBar behavior & add UserStoryService (#26)

* ccs improvement

* deleting wrong card

+++ src/app/components/party-user-stories-list/party-user-stories-list.component.ts
@@ -1,3 +1,4 @@
+import { LiteralArrayExpr } from '@angular/compiler';
 import { Component, OnInit, Output, EventEmitter, Renderer2, AfterViewInit, Input, OnChanges, SimpleChanges } from '@angular/core';
 import { NgToastService } from 'ng-angular-popup';
 import { UserStory } from 'src/app/models/user-story';
@@ -11,11 +12,14 @@ import { PartyService } from 'src/app/services/party.service';
 export class UserStoriesListComponent implements OnInit, AfterViewInit, OnChanges {
 
     @Input() partyID: string;
+    @Input() addedUserStory: UserStory;
     @Input() deletedUserStoryId: number;
     @Output() selectedUserStory = new EventEmitter<UserStory>();
+    @Output() resetSelectedUserStory = new EventEmitter<any>();
 
-    userStoriesList = new Array<any>();
+    userStoriesList = new Array<UserStory>()
     itemSelected: HTMLElement;
+    selectedUS: UserStory;
 
     constructor(private partyService: PartyService,
                 private render: Renderer2,
@@ -30,26 +34,32 @@ export class UserStoriesListComponent implements OnInit, AfterViewInit, OnChange
     }
 
     ngOnChanges(changes: SimpleChanges) {
-        console.log(changes);
-        if (!changes['deletedUserStoryId'].firstChange){
+        if ( changes['addedUserStory'] && (changes['addedUserStory'].previousValue != changes['addedUserStory'].currentValue) ) {
+            this.userStoriesList.push(this.addedUserStory);
+        }
+
+        if (changes['deletedUserStoryId'] && !changes['deletedUserStoryId'].firstChange){
             let usIDtoDelete = changes['deletedUserStoryId'].currentValue;
             this.deleteUSFromList(usIDtoDelete);
+            this.resetSelectedUS();
         }
     }
 
     asingClasses(){
         this.userStoriesList.forEach((us) => {
-            if (us.active == false) {
+            if (!us.isActive) {
                 let elementUS = document.getElementById(`us-${us.id}`);
                 this.render.addClass(elementUS, "disabled");
             }
-
         }); 
     }
 
     getPartyUserStories() {
         this.partyService.getPartyUserStories(this.partyID).subscribe((response) => {
-            this.userStoriesList = response; 
+            if(response)
+                this.userStoriesList = response; 
+            else
+                this.userStoriesList = new Array();
             
         },
             (apiError) => {
@@ -61,15 +71,22 @@ export class UserStoriesListComponent implements OnInit, AfterViewInit, OnChange
             });
     }
 
+    resetSelectedUS(){
+        this.resetSelectedUserStory.emit();
+        
+        if (this.itemSelected)
+            this.render.removeClass(this.itemSelected, "active");
+    }
+
     handleClickItem(us: UserStory){
         if(this.itemSelected)
             this.render.removeClass(this.itemSelected, "active");
 
+        this.selectedUS = us;
         this.itemSelected = document.getElementById(`us-${us.id}`);
         this.render.addClass(this.itemSelected, "active");
         
         this.selectedUserStory.emit(us);
-        
     }
 
     handleAddedUs(newUS){
